pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.49.0",
    "langchain-mcp-adapters>=0.0.3",
    "mcp>=1.6.0",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "ruff<1.0.0,>=0.7.3",
    "pyright<2.0.0,>=1.1.388",
    "pre-commit<5.0.0,>=4.0.1",
    "pytest>=8.3.5",
    "pytest-pretty>=1.2.0",
    "pytest-timeout>=2.3.1",
    "pytest-asyncio>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_client"]

[tool.uv]
default-groups = []

[tool.ruff]
line-length=100

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "ANN"]
ignore = [
    "ANN002", # untyped *args
    "ANN003", # untyped **kwargs
]

[tool.ruff.lint.per-file-ignores]
"**/tests/*" = [
    "ANN201",
]

[tool.pyright]
typeCheckingMode = "standard"
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
#log_cli = true
#log_cli_level = "DEBUG"
timeout = 300  # Global timeout for all tests (prevent indefinite hangs, but cancels all tests)
asyncio_mode = "auto"  # Automatically detects async test functions and fixtures and treats them as marked
asyncio_default_fixture_loop_scope = "session"
filterwarnings = [
    "error",
    # "ignore::DeprecationWarning:langchain_core",
    # "ignore::DeprecationWarning:langgraph",
    "ignore::DeprecationWarning:pydantic.v1.typing",

]
